dashboard/requirements-redesign.txt
Flask==2.3.3
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0  # For JWT tokens
Flask-CORS==4.0.0